+++ package.json
@@ -20,13 +20,13 @@
     "fastify": "^5.0.0",
     "neostandard": "^0.13.0",
     "tinybench": "^6.0.0",
-    "tsd": "^0.33.0"
+    "tstyche": "^7.0.0"
   },
   "scripts": {
     "lint": "eslint",
     "lint:fix": "eslint --fix",
     "test": "npm run test:unit && npm run test:typescript",
-    "test:typescript": "tsd",
+    "test:typescript": "tstyche",
     "test:unit": "node --test --experimental-test-coverage",
     "test:watch": "node --test --watch"
   },
